looking good

+++ modules/money_accounts/services.go
@@ -1,6 +1,7 @@
 package money_accounts
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/google/uuid"
@@ -11,7 +12,7 @@ import (
 
 func GetMoneyAccounts() []MoneyAccount {
 	var moneyAccounts []MoneyAccount
-	rows, err := database.DB.Query("SELECT * FROM money_accounts;")
+	rows, err := database.DB.Query("SELECT * FROM money_accounts WHERE id <> $1;", uuid.UUID{})
 	errors_handler.CheckError(err)
 	defer rows.Close()
 
@@ -37,9 +38,12 @@ func CreateMoneyAccount(fields MoneyAccountFields) (MoneyAccount, error) {
 	return nma, nil
 }
 
-func GetOneMoneyAccount(acount_id uuid.UUID) (MoneyAccount, error) {
+func GetOneMoneyAccount(account_id uuid.UUID) (MoneyAccount, error) {
 	var ma MoneyAccount
-	row := database.DB.QueryRow("SELECT * FROM money_accounts WHERE id = $1;", acount_id)
+	if account_id == (uuid.UUID{}) {
+		return ma, fmt.Errorf(errors_handler.DB001)
+	}
+	row := database.DB.QueryRow("SELECT * FROM money_accounts WHERE id = $1;", account_id)
 	err := row.Scan(&ma.ID, &ma.Name, &ma.Balance, &ma.Details, &ma.Currency, &ma.CreatedAt, &ma.UpdatedAt)
 	if err != nil {
 		if errors_handler.CheckEmptyRowError(err) {
@@ -52,6 +56,9 @@ func GetOneMoneyAccount(acount_id uuid.UUID) (MoneyAccount, error) {
 
 func UpdateMoneyAccount(account_id uuid.UUID, fields MoneyAccountFields) (MoneyAccount, error) {
 	var uma MoneyAccount
+	if account_id == (uuid.UUID{}) {
+		return uma, fmt.Errorf(errors_handler.DB001)
+	}
 	row := database.DB.QueryRow("UPDATE money_accounts SET name = $1, details = $2, currency = $3, updated_at = $4 WHERE id = $5 RETURNING *;",
 		fields.Name, fields.Details, fields.Currency, time.Now(), account_id)
 	err := row.Scan(&uma.ID, &uma.Name, &uma.Balance, &uma.Details, &uma.Currency, &uma.CreatedAt, &uma.UpdatedAt)
@@ -64,9 +71,12 @@ func UpdateMoneyAccount(account_id uuid.UUID, fields MoneyAccountFields) (MoneyA
 	return uma, nil
 }
 
-func getAccountsName(acount_id uuid.UUID) (string, error) {
+func getAccountsName(account_id uuid.UUID) (string, error) {
 	name := ""
-	row := database.DB.QueryRow("SELECT name FROM money_accounts WHERE id = $1;", acount_id)
+	if account_id == (uuid.UUID{}) {
+		return name, fmt.Errorf(errors_handler.DB001)
+	}
+	row := database.DB.QueryRow("SELECT name FROM money_accounts WHERE id = $1;", account_id)
 	err := row.Scan(&name)
 	if err != nil {
 		if errors_handler.CheckEmptyRowError(err) {
@@ -79,6 +89,9 @@ func getAccountsName(acount_id uuid.UUID) (string, error) {
 
 func DeleteOneMoneyAccount(account_id uuid.UUID) (common.ID, error) {
 	id := common.ID{}
+	if account_id == (uuid.UUID{}) {
+		return id, fmt.Errorf(errors_handler.DB001)
+	}
 	row := database.DB.QueryRow("DELETE FROM money_accounts WHERE id = $1 RETURNING id;", account_id)
 	err := row.Scan(&id.ID)
 	if err != nil {
@@ -99,6 +112,9 @@ func ResetAccountsBalance(account_id uuid.UUID) (common.ID, error) {
 
 func setAccountsBalance(account_id uuid.UUID, balance float64) (common.ID, error) {
 	id := common.ID{}
+	if account_id == (uuid.UUID{}) {
+		return id, fmt.Errorf(errors_handler.DB001)
+	}
 	row := database.DB.QueryRow("UPDATE money_accounts SET balance = $1 WHERE id = $2 RETURNING id;",
 		balance, account_id)
 	err := row.Scan(&id.ID)
@@ -112,5 +128,5 @@ func setAccountsBalance(account_id uuid.UUID, balance float64) (common.ID, error
 }
 
 func DeleteAllMoneyAccounts() {
-	database.DB.QueryRow("DELETE FROM money_accounts;")
+	database.DB.QueryRow("DELETE FROM money_accounts WHERE id <> $1;", uuid.UUID{})
 }
